requirements.txt
absl-py==2.1.0
numpy==2.0.1
scipy==1.14.0
wget==3.2
